backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.7.4
httpx==0.27.2
python-dotenv==1.0.1
yfinance==0.2.41
finnhub-python==2.4.20
apscheduler==3.10.4
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
openai==1.44.1
requests==2.32.3
pycoingecko==3.1.0
